Flatten config so each config is one artifact.
Previously a given config could contain multiple artifacts but this
caused lots of extra complexity around aggregating the results of the
artifacts into a single set of outputs.

It's conceptually simpler to just say that each config represents one
standalone artifact, which then maps more closely onto the underlying idea
of a root terraform module that declares a bunch of resources and
produces some outputs.

This also has the nice side-effect of allowing artifact definitions to
span across multiple config files, which wasn't possible when artifact
was a block construct in the HCL.