Completed proof-of-concept of the overall build workflow.
Here we are able to create all of the resources in a config and then
split the state into temporary and result parts. We can then (as a
separate step) clean up the temporary parts, leaving just the result
parts.

For some reason at present the outputs in the config don't make it into
the resulting state, which prevents the result from being truly useful.